--- src/test/kotlin/logic/useCases/mealSearchUseCase/KmpMatcherTest.kt
@@ -1,7 +1,0 @@
-package logic.useCases.mealSearchUseCase
-
-import org.junit.jupiter.api.Assertions.*
-
-class KmpMatcherTest{
-
-}
